requirements.txt
fastapi
numpy
pyzbar
pillow
uvicorn[standard]